requirements.txt
click
psutil
google-api-python-client
oauth2client
six
requests
bs4
beautifulsoup4
httplib2
PyYAML
